feat(@vates/types): expose XoApp
