+++ @vates/types/package.json
@@ -9,6 +9,7 @@
       "types": "./dist/index.d.mts"
     },
     "./xo": "./dist/xo.d.mts",
+    "./xo-app": "./dist/xo-app.d.mts",
     "./xen-api": "./dist/xen-api.d.mts",
     "./common": {
       "default": "./dist/common.mjs",
