+++ @vates/types/src/index.mts
@@ -1,5 +1,6 @@
 export * from './common.mjs'
 export * from './xen-api.mjs'
 export * from './xo.mjs'
+export * from './xo-app.mjs'
 export * from './lib/xen-orchestra-xapi.mjs'
 export * from './lib/complex-matcher.mjs'
