+++ CHANGELOG.unreleased.md
@@ -34,6 +34,7 @@
 
 <!--packages-start-->
 
+- @vates/types minor
 - @xen-orchestra/web minor
 - @xen-orchestra/web-core minor
 - @xen-orchestra/xapi patch
